No more conversion warnings

+++ src/Matrix.cpp
@@ -58,25 +58,25 @@ bool Matrix::getInvert(matrix_t out)
 
     r0[0] = MAT(m,0,0), r0[1] = MAT(m,0,1),
         r0[2] = MAT(m,0,2), r0[3] = MAT(m,0,3),
-        r0[4] = 1.0, r0[5] = r0[6] = r0[7] = 0.0,
+        r0[4] = 1.0f, r0[5] = r0[6] = r0[7] = 0.0f,
 
         r1[0] = MAT(m,1,0), r1[1] = MAT(m,1,1),
         r1[2] = MAT(m,1,2), r1[3] = MAT(m,1,3),
-        r1[5] = 1.0, r1[4] = r1[6] = r1[7] = 0.0,
+        r1[5] = 1.0f, r1[4] = r1[6] = r1[7] = 0.0f,
 
         r2[0] = MAT(m,2,0), r2[1] = MAT(m,2,1),
         r2[2] = MAT(m,2,2), r2[3] = MAT(m,2,3),
-        r2[6] = 1.0, r2[4] = r2[5] = r2[7] = 0.0,
+        r2[6] = 1.0f, r2[4] = r2[5] = r2[7] = 0.0f,
 
         r3[0] = MAT(m,3,0), r3[1] = MAT(m,3,1),
         r3[2] = MAT(m,3,2), r3[3] = MAT(m,3,3),
-        r3[7] = 1.0, r3[4] = r3[5] = r3[6] = 0.0;
+        r3[7] = 1.0f, r3[4] = r3[5] = r3[6] = 0.0f;
 
     /* choose pivot - or die */
     if (fabs(r3[0])>fabs(r2[0])) SWAP_ROWS(r3, r2);
     if (fabs(r2[0])>fabs(r1[0])) SWAP_ROWS(r2, r1);
     if (fabs(r1[0])>fabs(r0[0])) SWAP_ROWS(r1, r0);
-    if (0.0 == r0[0])  return false;
+    if (0.0f == r0[0])  return false;
 
     /* eliminate first variable     */
     m1 = r1[0]/r0[0]; m2 = r2[0]/r0[0]; m3 = r3[0]/r0[0];
@@ -84,31 +84,31 @@ bool Matrix::getInvert(matrix_t out)
     s = r0[2]; r1[2] -= m1 * s; r2[2] -= m2 * s; r3[2] -= m3 * s;
     s = r0[3]; r1[3] -= m1 * s; r2[3] -= m2 * s; r3[3] -= m3 * s;
     s = r0[4];
-    if (s != 0.0) { r1[4] -= m1 * s; r2[4] -= m2 * s; r3[4] -= m3 * s; }
+    if (s != 0.0f) { r1[4] -= m1 * s; r2[4] -= m2 * s; r3[4] -= m3 * s; }
     s = r0[5];
-    if (s != 0.0) { r1[5] -= m1 * s; r2[5] -= m2 * s; r3[5] -= m3 * s; }
+    if (s != 0.0f) { r1[5] -= m1 * s; r2[5] -= m2 * s; r3[5] -= m3 * s; }
     s = r0[6];
-    if (s != 0.0) { r1[6] -= m1 * s; r2[6] -= m2 * s; r3[6] -= m3 * s; }
+    if (s != 0.0f) { r1[6] -= m1 * s; r2[6] -= m2 * s; r3[6] -= m3 * s; }
     s = r0[7];
-    if (s != 0.0) { r1[7] -= m1 * s; r2[7] -= m2 * s; r3[7] -= m3 * s; }
+    if (s != 0.0f) { r1[7] -= m1 * s; r2[7] -= m2 * s; r3[7] -= m3 * s; }
 
     /* choose pivot - or die */
     if (fabs(r3[1])>fabs(r2[1])) SWAP_ROWS(r3, r2);
     if (fabs(r2[1])>fabs(r1[1])) SWAP_ROWS(r2, r1);
-    if (0.0 == r1[1])  return false;
+    if (0.0f == r1[1])  return false;
 
     /* eliminate second variable */
     m2 = r2[1]/r1[1]; m3 = r3[1]/r1[1];
     r2[2] -= m2 * r1[2]; r3[2] -= m3 * r1[2];
     r2[3] -= m2 * r1[3]; r3[3] -= m3 * r1[3];
-    s = r1[4]; if (0.0 != s) { r2[4] -= m2 * s; r3[4] -= m3 * s; }
-    s = r1[5]; if (0.0 != s) { r2[5] -= m2 * s; r3[5] -= m3 * s; }
-    s = r1[6]; if (0.0 != s) { r2[6] -= m2 * s; r3[6] -= m3 * s; }
-    s = r1[7]; if (0.0 != s) { r2[7] -= m2 * s; r3[7] -= m3 * s; }
+    s = r1[4]; if (0.0f != s) { r2[4] -= m2 * s; r3[4] -= m3 * s; }
+    s = r1[5]; if (0.0f != s) { r2[5] -= m2 * s; r3[5] -= m3 * s; }
+    s = r1[6]; if (0.0f != s) { r2[6] -= m2 * s; r3[6] -= m3 * s; }
+    s = r1[7]; if (0.0f != s) { r2[7] -= m2 * s; r3[7] -= m3 * s; }
 
     /* choose pivot - or die */
     if (fabs(r3[2])>fabs(r2[2])) SWAP_ROWS(r3, r2);
-    if (0.0 == r2[2])  return false;
+    if (0.0f == r2[2])  return false;
 
     /* eliminate third variable */
     m3 = r3[2]/r2[2];
@@ -117,13 +117,13 @@ bool Matrix::getInvert(matrix_t out)
         r3[7] -= m3 * r2[7];
 
     /* last check */
-    if (0.0 == r3[3]) return false;
+    if (0.0f == r3[3]) return false;
 
-    s = 1.0/r3[3];              /* now back substitute row 3 */
+    s = 1.0f/r3[3];              /* now back substitute row 3 */
     r3[4] *= s; r3[5] *= s; r3[6] *= s; r3[7] *= s;
 
     m2 = r2[3];                 /* now back substitute row 2 */
-    s  = 1.0/r2[2];
+    s  = 1.0f/r2[2];
     r2[4] = s * (r2[4] - r3[4] * m2), r2[5] = s * (r2[5] - r3[5] * m2),
         r2[6] = s * (r2[6] - r3[6] * m2), r2[7] = s * (r2[7] - r3[7] * m2);
     m1 = r1[3];
@@ -134,7 +134,7 @@ bool Matrix::getInvert(matrix_t out)
         r0[6] -= r3[6] * m0, r0[7] -= r3[7] * m0;
 
     m1 = r1[2];                 /* now back substitute row 1 */
-    s  = 1.0/r1[1];
+    s  = 1.0f/r1[1];
     r1[4] = s * (r1[4] - r2[4] * m1), r1[5] = s * (r1[5] - r2[5] * m1),
         r1[6] = s * (r1[6] - r2[6] * m1), r1[7] = s * (r1[7] - r2[7] * m1);
     m0 = r0[2];
@@ -142,7 +142,7 @@ bool Matrix::getInvert(matrix_t out)
         r0[6] -= r2[6] * m0, r0[7] -= r2[7] * m0;
 
     m0 = r0[1];                 /* now back substitute row 0 */
-    s  = 1.0/r0[0];
+    s  = 1.0f/r0[0];
     r0[4] = s * (r0[4] - r1[4] * m0), r0[5] = s * (r0[5] - r1[5] * m0),
         r0[6] = s * (r0[6] - r1[6] * m0), r0[7] = s * (r0[7] - r1[7] * m0);
 
@@ -306,17 +306,17 @@ void Matrix::rotate(vec_t ax, vec_t ay, vec_t az)
     matrix_t xmat, ymat, zmat, tmp, tmp2;
 
     xmat[ 0]=1;        xmat[ 1]=0;        xmat[ 2]=0;        xmat[ 3]=0;
-    xmat[ 4]=0;        xmat[ 5]=cos(ax);  xmat[ 6]=sin(ax);  xmat[ 7]=0;
-    xmat[ 8]=0;        xmat[ 9]=-sin(ax); xmat[10]=cos(ax);  xmat[11]=0;
+    xmat[ 4]=0;        xmat[ 5]=cosf(ax);  xmat[ 6]=sinf(ax);  xmat[ 7]=0;
+    xmat[ 8]=0;        xmat[ 9]=-sinf(ax); xmat[10]=cosf(ax);  xmat[11]=0;
     xmat[12]=0;        xmat[13]=0;        xmat[14]=0;        xmat[15]=1;
 
-    ymat[ 0]=cos(ay);  ymat[ 1]=0;        ymat[ 2]=-sin(ay); ymat[ 3]=0;
+    ymat[ 0]=cosf(ay);  ymat[ 1]=0;        ymat[ 2]=-sinf(ay); ymat[ 3]=0;
     ymat[ 4]=0;        ymat[ 5]=1;        ymat[ 6]=0;        ymat[ 7]=0;
-    ymat[ 8]=sin(ay);  ymat[ 9]=0;        ymat[10]=cos(ay);  ymat[11]=0;
+    ymat[ 8]=sinf(ay);  ymat[ 9]=0;        ymat[10]=cosf(ay);  ymat[11]=0;
     ymat[12]=0;        ymat[13]=0;        ymat[14]=0;        ymat[15]=1;
 
-    zmat[ 0]=cos(az);  zmat[ 1]=sin(az);  zmat[ 2]=0;        zmat[ 3]=0;
-    zmat[ 4]=-sin(az); zmat[ 5]=cos(az);  zmat[ 6]=0;        zmat[ 7]=0;
+    zmat[ 0]=cosf(az);  zmat[ 1]=sinf(az);  zmat[ 2]=0;        zmat[ 3]=0;
+    zmat[ 4]=-sinf(az); zmat[ 5]=cosf(az);  zmat[ 6]=0;        zmat[ 7]=0;
     zmat[ 8]=0;        zmat[ 9]=0;        zmat[10]=1;        zmat[11]=0;
     zmat[12]=0;        zmat[13]=0;        zmat[14]=0;        zmat[15]=1;
